requirements.txt
fastapi
httpx
uvicorn[standard]
paho-mqtt
requests
